Renamed routine to avoid confusion about what it does.
The affected routine ("Total_Area_Bisection_Along_Plane") does NOT do full volumetric sub-segmentation.
It operates on 2D planar slices embedded in 3D space. If your planar objects are associated with a non-
zero 'thickness' in 3D space (making "slabs"), and the slabs do not overlap, then this routine will
compute a discrete approximation of a volumetric subsegmentation via bisection. Only an approximation
because the objects are genuinely 2D planar objects! In particular, the slabs will not be sliced
properly.

If your 3D object is either (1) composed of many slabs OR (2) being sliced by a plane that is oblique
or perpendicular (i.e., just NOT nearly parallel) to the plane the slabs are defined in then the
approximation should generally be of acceptable quality.