Replaced bisection bounding plane generation.
The existing implementation used planes, which involves vector distances and thus a square and square-root operation,
which can cause numerical losses that can potentially be significant. I replaced it with a simple dot-product
implementation that appears to be more robust. The user can still override the initial bounding planes if the default is
not suitable.