Added support for voxel adjacency in Mutate_Voxels().
It was plumbed in but not actually implemented until I could figure out how best to handle boundary edges.
The selected method (only using neughbours if they exist) could potentially cause problems depending on the
aggregate function being used, but I can't imagine a scenario in which a non-summarizing aggregate would
need to be paired with accurate boundary extrapolation or similar.