Mutate_Voxels: give voxel visitors access to image.
Previously the only way voxel visitor functions could be made aware of which image they were operating on was to capture
the image as a closure parameter. They also needed to modify the images in-place rather than use a working image. This
presented issues for pre-packaged or delegated voxel visitors. This commit adds a function parameter filled with a
reference to the image being processed, which can be used to introspect the image to, e.g., determine voxel coordinates
or check other channels or adjacent voxels.