Added rudimentary surface mesh class and IO routines.
This class is 'low-tech'. More sophisticated implementations were becoming difficult to manage, causing CI tests to run
for an hour and present roadblocks for simple interchange. The Ygor fv_surface_mesh class was added to simplify
interchange and provide a simple lowest-common-denominator format for DCMA. A more sophisticated implementation with
attributes might be added (or retrofitted) if necessary. Probably a half-edge implementation. But at the moment it is
not necessary.