samples_1D: added support for metadata stringification and serialization.
Metadata now round-trips when samples_1D's are stringified (i.e., in preparation for plotting) or serialized (i.e., in
preparation for passing over a network or some other interface). Metadata is base-64 encoded to deal with problematic
strings, but only non-printable characters are encoded when stringified. This permits easier interoperability, or at
least easier keyword grepping.