Provide a compile-time definitions file.
This will ensure that users of the library do not need to supply the same conditional compilation flags that were
supplied during compilation of the library; they are now baked into the headers using CMake configure_file().
There is virtually no other value in this arrangement. In fact it may increase build times when the configuration
changes, though this is a minor inconvenience since libygor will almost certainly need to be recompiled from scratch
anyways.