Downgrade const requirement.
The requirement to make all arguments const ref wrappers will require cascading changes downstream. While it is clearer
to the caller when arguments are explicitly const, ultimately in this case no changes are made to the caller's data but
it is not reflected in the signature. Note that internally all refw's are casted to const using cref anyways.