planar_image_adjacency: switch to linear index lookup.
Was previously using a map-based lookup, which is only needed for the converse lookup (image-to-index).
While with a map it was sensible to have the indices offset by a fixed number to avoid collision with the map's default
index (0), for consistency with image row and column numbers, and to facilitate easier enumeration of all images (i.e.,
start at zero, iterate positively until the index can not be found), this offset was removed. The first image will now
be guaranteed index 0. The user must be careful not to create erroneous references to it via the deafult constructor
though.

A test-case was also created to verify the modified implementation.