Mutate_Voxels: gracefully handle disconnected regions.
The new fast algorithm liberally used std::abs() to ensure projections had a positive sign. Also the (0,0) voxel
(centre) position was used as a positioning reference. However, whole-image contours will necessarily extend beyond the
centres of all voxel, and the std::abs()'s caused the zeroth row and column to appear to be outside the bounding
contour. This commit addresses the issue. It also adds an explicit check that the simplistic rows/columns-to-visit
subroutine did not miss a valid row/column, and also adjusts the logic in a related routine to be more clear about how
rows/columns-to-visit short-circuits.