Revamped vec3 stringification.
The existing stringification relied on the input stream state being checked by the user, but the stream wasn't always
made available to be checked. This commit replaces the basic stringstream interface with a std::stod-based approach that
appears to better handle very large numbers (e.g., the max finite double).

A more concise approach may be taken in the future, but for the time being this routine satisfies the included test
cases which likely cover realistic, 'physical' doubles that would be encountered normally (i.e., spanning -1E-308 to
+1E308 without writing an excessive number of decimal digits).