Added minimal matrix class.
The intention of this class is a sort of 'polyfill' -- to provide minimal matrix support when something like Eigen is
not available. This implementation was specifically created to support a thin plate spline implementation,
wherein serializability, portability, API and ABI stability, and template compilation times need to be controlled.
Performance of this implementation will likely be abysmal compared to any legitmate matrix handling library.

This implementation is also written with Eigen interoperability in mind, though no adaptators have been written yet.