Avoid altering src/ for typical builds.
This commit switches where the generated file YgorDefinitions.h is placed. It was previously placed into src/ beside
YgorDefinitions.h.in, but this is slightly annoying for the typical build process. Note: it doesn't impact the 'clone'
method I normally use (via rsync'ing the entire repo) but does impact the usual 'cd build ; cmake ../ ; make' workflow
since it pollutes *outside* the build directory. The new style is cleaner in terms of file layout, but slightly more
complicated in the CMake script, and also requires an additional '-I' to be passed to the compiler.