YgorFilesDirs: made more directory routines portable.
The directory listing functions may change slightly in this commit, specifically the presence of the preceeding
directory. Where earlier routines may have prepended the base dir argument to the output, the updated routines merely
return the (full) canonical path. In most cases this should not impact users, but cases where the full path is not
expected and was added manually afterward will result in invalid paths. A workaround for this is to rely on dedicated
pathname component extractors, or switch to std::filesystem. I wasn't able to find any instances where this could occur,
but my search was not exhaustive.

The test cases have also been expanded to explicitly verify the files and directories listed are as expected.