Reimplement 'unique' metadata selectors.
The original implementation selected 'unique' values, i.e., only metadata values where there was a single matching key.
While this is indeed technically what the name suggests, it wasn't helpful. And it wasn't widely used (and there are
indications that where it was used, there was unexpected behaviour).

The 'unique' metadata selector routines were replaced with 'distinct' selector routines that are more inline with the
`uniq` GNU program, and much more usable.