contour_collection: added metadata selector.
The typical use-case for metadata selection is to assume that all contours in a collection have mostly the same
(if not identical) key-value pairs. This commit adds a convenience routine that processess all key-value pairs, finds
the statistical (categorical) mode, and returns it as a representative of the whole contour_collection.

A minimal test case is also provided for all the contour_collection metadata selector routines.