point_set: add support for (optional) normals.
Whether normals is used or not is up to the end user, and they are not managed. In particular, there are currently no
routines that are 'clever' enough to do anything with normals, especially if vertices are altered in any way.

If normals are present, the number of normals and vertices must match when reading or writing OBJ or OFF files.