Added vec{2,3} class layout checks.
Specifically, these tests can be used to monitor that no extra members have been added, no padding is inserted, and
arrays of vec{2,3}'s can be accessed at the individual coordinate level, i.e., an array of two vec3<double> can be
reinterpret_cast'ed to an array with 6 doubles. This is especially important for interoperation with C libraries and for
code that must be performant in passing data around (e.g., OpenGL, OpenCL, SYCL, etc).