ASCII PLY IO routines: support arbitrary metadata.
Metadata elements are encoded and decoded for ASCII PLY IO routines. Data is written as a PLY comment and is
automatically base64 encoded if the content is expected to interfere with the PLY structure (unprintable and special
printable characters). Base64 decoding is automatic. Note that, whenever possible, metadata is left raw to maximize
interoperability and editability/injection.