PLY: support reading and writing binary files.
This implementation has some caveats: the stream should be opened in the appropriate mode (text) even if a binary file
is expected, and also I don't think the current implementation will work at all on Windows when CRLF line endings are
present. There is also risk that some files will happen to contain a CRLF by chance (in the vertex/face data), which
will erroneously cause the file to fail to be loaded. I don't think data loss can occur, since the file will fail to be
readable. Unix systems that only use '\n' should be fine.

A fix is incoming, but it might not actually work, depending on whether I can wrestly a file descriptor/number from the
stream, and then successfully use implementation-specific routines to convert the stream behaviour in-place. It's
probably best to just open PLY files in binary mode on all systems, and I might resort to that instead.