PLY: be more stringent with header magic number.
The use of getline for the 'ply' signature means that files without line endings (e.g., large binary files) will be read
entirely in the first std::getline(). Performance in this case will be terrible. Not all formats *require* the signature
to be present, but PLY seems to. So we can quickly reject non-PLY files without risk of paging in the whole file.

Ideally parser leniency could be controlled with a 'strictness' parameter, but for the time being this seems like a
safer choice.