Refactor endianness checks.
Replaced runtime checks with constexpr compile-time checks where appropriate. I had hoped to use std::endian for this,
but the feature-test macros required to use it (currently) were too gnarly. It was easier to just rely on compiler
definitions. The fallback (little-endian) is probably >99% of machines this code will ever run on, and this
functionality may not be important in those cases anyways.