PLY: increase portability.
Previously, the explicit size of the facet index was written as-is. For non-32 bit integers, this caused problems with
PLY parsers that do not handle non-32 bit integers. Additionally, use of 64 bit integers is convenient at runtime, but
for binary storage it can be wasteful if there are insufficient vertices to warrant 64 bit indices.

This commit addresses both issues (portability and storage space) when using 64 bit indices. When the number of vertices
is less than (2^32)-1, the integer size is promoted/converted to 32 bit. (Even if smaller indices are used in-memory!)
The smallest integer written will be 32 bit, which is the standard PLY integer size for 'int's. A larger size will only
be used if required, to avoid data loss.

So portability is improved until it would cause data loss, and there is less data written in the typical case (64 bit
indices with a small number of faces present).