PLY: prefer signed ints for face vert indices.
This improves portability since previously unsigned ints were written, but were being advertised as signed ints.

I hoped that this would improve the odds of Meshlab accepting binary PLYs, but no luck. The source at
https://github.com/cnr-isti-vclab/vcglib/blob/f7871eed2d54bee704b9f66dc538c921602323cb/wrap/io_trimesh/import_ply.h
indicates (recent versions of) Meshlab *should* accept 8, 16, and 32 bit face indices, but the release v2020.08 (built
on 11 Aug 2020) only seems to support int32 for a binary PLY with 33'000 vertices. Hopefully this will be fixed in a
future release. If so, the adaptive type-shrink in Ygor's PLY export function can be switched on for 8 and 16 bit ints,
and maybe expanded to include 8 and 16 bit uints.