Added routine to find convex hull.
I'm not terribly happy about the interface, but I wanted to make it usable for multiple types of inputs (e.g., multiple
polygons, point clouds, surface meshes, etc.). I think I might switch to an iterator-based interface in the future.
Minimal vertex access is needed, including no read-ahead. At any given iteration, only vertices on the hull and a single
candidate vertex are ever needed. Also, vertices that are dropped from the hull are never returned, So I might be able
to keep access to a single pass and also keep memory usage limited to the hull size.

I'm also not terribly happy with the need to update the complete face adjacency index after every hull addition. This
info can largely be cached and re-used. I thought it would be best to get a feel for how this approach scales though.
10'000 randomly-distributed vertices takes a few seconds, so some benchmarking is needed.

This commit also adds a few rudimentary tests. They basically only check that exceptions are not thrown, and only
minimally confirm the hull has the correct number of faces. Again, I didn't want to make too many tests with the
interface in flux.