Convex_Hull: incrementally build adjacency list.
Previously, the adjacency list was continuously regenerated afresh everytime the hull was modified. This resulted in
cubic scaling. This commit instead builds the adjacency list incrementally, which allows for 1-2 orders of magnitude
more points to be processed in the same amount of time (from 10'000 to 100'000 - 1'000'000 based on random points inside
the unit cube with a slightly larger cube -- see tests). Note that performance on real-world (or pathological) data will
likely be considerably slower than this artificial setup.