Convex_Hull: add robust seed tet finder.
The prior implementation was not able to handle degenerate vertices when building the seed tetrahedron. This commit adds
an adaptive check that ensures vertices are not congruent, add finite area, and add finite volume.

Runtime may be impacted for pathological cases, such as when there are N+1 vertices in total and the first N are
co-planar, which is not unexpected for things like planar contour stacks. However, this approach will successfully
navigate most degeneracies. Degeneracies that involve numerical imprecision may not be properly handled at this point,
though vertex-vertex coincidence is explicitly checked.

Test cases with commons kinds of degeneracies were added.