ConvexHull: replace face vector with map.
Faces continually pop into and out of existence. With a vector, it's cumbersome to garbage collect deleted faces.
Especially so when the adjacency list would also need to be (fully) updated. Using a map helps keeps the face numbers
intact while allowing old faces to be reaped on-the-fly.

The penalty for key-value lookup seems to be on par with vector memory allocation costs, so there isn't much speed up in
this commit. However, it does make it easier to potentially implement divide-and-conquer since some of the most awkward
boilerplate is now cleaned up.

Note that this commit has not been optimized. There are still many performace gains to be had (after the horrendous
scaling is taken care of).