simplify_inner_triangles: reject candidate faces with negative orientation.
This commit removes previous behaviour whereby the final candidate triangle in any patch was allowed to have a negative
orientation, so long as it satisfied all other criteria. Previously, the face was simply reoriented to be consistent
with the other faces, but would sometimes result in replacement faces folding back into itself (pinching).

This commit also adds a user-specified criteria to either increase or decrease the face alignment criteria. Previously
faces needed to point in the same direction (as the first face). Now, faces need to align within the user-provided angle
with the area-weighted surface patch average normal.

Note that this approach is not fool-proof, since it can still cause faces to pinch and fold back into other faces. But
it is much less likely for meshes with high-quality triangles of roughly the same size.