simplify_inner_triangles: add more stringent checks.
This commit adds an additional alignment constraint on the patch: all original faces have to have the same orientation
as the area-weighted average normal in order to proceed with simplification.

This commit also adds checks for degeneracies and simplifications that would result in surface pinches. Such pinches can
occur when the input surface is a tetrahedron or have a triangular cross-section.

Note that this commit eliminates non-manifold simplifications that appeared during a test with a complicated, real-world
marching cubes mesh from a highly non-convex shape.