FITS: move to FITSv4.0 spec, add full metadata support.
The FITS format works well as a simple interchange format, but was lacking support for image key-value metadata. This
commit adds support for both reading and writing metadata including transparent escaping (quoting, base64).

This commit also updates the writer to conform to the FITS4.0 specification. It now emits keywords using '=' syntax and
strings. Integers and floats are now fixed-width. Long strings are supported using the '&' and CONTINUE syntax.

Note that the reader continues to be picky about Ygor-specific keywords, requiring them to initialize the spatial aspect
of the image. While this may be relaxed in the future, the current rationale is to simplify interchange between users of
Ygor.