FITS: be explicit about FITS extensions.
The writer now emits an explicit statement that no extensions follow, and the reader explicitly checks for extensions.
At the moment the reader doesn't do anything with the information, but it will help support the future IMAGE extension
for receiving a collection of images. This will address a major pain point (using multiple files to store a logically
connected image array).