FITS: support multi-image FITS files.
This commit replaces the prior single-image FITS reader and writer with multi-image capable reader and writer.

The writer will still emit a single image FITS file when only a single image is present, but will correctly 'upgrade' to
multi-image if additional images are present.

The reader will attempt to read all images into a planar_image_collection rather than the previous (a single
planar_image). The reader is not fully conformant -- notably, some functionality is not supported (e.g., the INHERIT
keyword, extensions other than IMAGE, and having a primary header with no associated data).

However, the reader is capable of reading files emitted by the writer, and fitsverify was used to confirm that the
writer produces files according to the FITSv4.0 specification + IMAGE extension.