FITS: workaround for excess precision.
This commit provides a workaround for a limitation encountered when losslessly encoding lengthy stringified
vec3<double>. The FITS long string continuation is now used for all YGOR metadata that are reasonably expected to exceed
the 70 character limit, including the column and row unit vectors which can easily exceed the limit, e.g.,

YGORANKR= '(0, 0, 0)'
YGOROFST= '(-129.73197618250001, -144.45973882281999, -42.354786517801998)'
YGORROWU= '(3.1147593999999945e-08, 0.99278502328327822, -0.11990787106949979)'
YGORCOLU= '(0.99997773673050938, -0.00080015021889999949, -0.00662463606209999&'
CONTINUE  '59)'

The 'YGORROWU' metadata is exactly 80 characters long, and the 'YGORCOLU' metadata slightly exceeds 80 characters.
Previously, attempting to export this image would fail because there was insufficient room and the truncation would be lossy and malformed.