contours: provide Get_Unsigned_Area() for contours and contour collections.
This routine is needed for correct area calculation with multi-orientation contour collections.
Signed area is needed to correctly calculate individual contour areas, which allows non-convex polygons to be handled.
However, when looking for the total unsigned area for multiple contours, the individual contour area is not relevant.
One solution is to reorient all contours to be uniform. Another is to provide an unsigned area calculation. Now both
options are available in Ygor.