Total_Area_Bisection_Along_Plane: use unsigned areas.
This commit switches from use of signed areas to unsigned areas, which makes it more robust to non-uniform contour
orientation and (somewhat) to non-planarity, which is important when contours are rotated off the cardinal axes due to
numerical imprecision.

This commit also retains a 'current best' plane in case bisection overshoots the optimal split but doesn't have enough
iterations to reach the optimal. Previously, the final result at the final iteration was taken which can easily result
in clipped extrema and small gaps between sub-segments. (Note: gaps and overlap are still possible -- the fractional
area tolerance is used to reduce overlap/gaps to an acceptable level.)