YgorLog: catch exceptions from callbacks.
Not sure how to handle callbacks that throw inside the global logging object. For now, I will silently catch exceptions
and continue on. Not sure what else can be done except immediately terminate, but that might be at odds with whatever we
want to log (i.e., try to capture as much info as possible before termination).