YgorLog: add opt-in terminal synchronization mutex.
Users of Ygor can opt-in to using this shared global mutex to synchronize *lines* emitted to a terminal. The YgorLog
emitter opt-in to this explicitly. Otherwsie, if users do not opt-in then at worst terminal output will be jumbled.
(Characters are written atomically, so there is a data race but no undefined behaviour as far as I can tell.)