YgorImages: re-define row and column orientation unit vectors.
Previously, the implementation used a confusing and inconsistent blend of interpretations where the row and column unit
vectors would either refer to the direction *along* a row/column, or *between* rows/columns. Part of this inconsistency
was the interpretation of the pixel width and height, which were frequently swapped, and correct placement of the row
and column during indexing.

Note that the rationale behind the previous implementation was to try keep the implementation flexible and agnostic to
the specific way that pixels/voxels were encoded. However, since the implementation now provides export to a number of
formats (where the conventions *do* matter), the implementation is now being harmonized.

This commit swaps the implementation to the convention where row_unit and col_unit define the direction *along* a
row/column. There is still sufficient opportunity for confusion going forward due to the name 'row_unit' and 'col_unit',
especially when combined with row and column counts for spatial positioning. So here is a summary guide:

|  Interpretation and Conventions:
|      top-left corner of an image: row = 0, column = 0 voxel
|      bottom-left corner of an image: row = (number_of_rows - 1), column = 0 voxel
|      top-right corner of an image: row = 0, column = (number_of_columns - 1) voxel
|
|      image anchor + image_offset = 3D position of the centre of the row = 0, column = 0 voxel
|
|      row_unit = unit vector *along* a row
|      col_unit = unit vector *along* a column
|
|      pxl_dx = voxel width
|      pxl_dy = voxel height
|      pxl_dz = voxel thickness / spatial depth (when viewed as a 2D image)
|
|  Adjacent voxels:
|      R' = R + row_unit * pxl_dx  <--- translating along the same row, but different columns
|      R' = R + col_unit * pxl_dy  <--- translating along the same column, but different rows
|
|  Indexing:
|      R' = R + row_unit * pxl_dx * column_number  <--- translates to the given column number
|             + col_unit * pxl_dy * row_number     <--- translates to the given row number
|
|  Constructors:
|      planar_image<>.init_spatial(dx, dy, dz, ...) = planar_image<>.init_spatial(width, height, dz)
|
|      DICOM 'PixelSpacing' = 'dy/dx' = height/width
|
|  Image dimensions:
|      pxl_dx * number_of_columns = width of an image
|      pxl_dy * number_of_rows    = height of an image
|      pxl_dz                     = thickness of an image
|
|  Forbidden combinations, which do not make sense:
|      pxl_dx * row_number
|      pxl_dy * column_number
|
|      row_unit * row_number
|      col_unit * column_number
|
|      row_unit * pxl_dy
|      col_unit * pxl_dx