Convex_Hull: attempt to truncate inputs to improve numerical stability.
This commit truncates all vertices to a maximum of 2 digits of decimal precision with the aim to gauge whether numerical
stability improves. The hull algoorithm is being used for some pathological cases for which this commit *might* help
improve stability. But it's hard to test in situ, so this commit will be reverted if there is no improvement...